--- src/base/static/components/organisms/main-map.scss
@@ -1,4 +1,0 @@
-.mapseed__map-control {
-  font-family: FontAwesome;
-  font-size: 1.4em;
-}
